Quickstarter Introduction.
This documentation makes it crystal clear how to create a quickstarter
It also shows how to create one from my own website. 
I am not yet sure what the required commands are, I just put in three dots ...
 I am sure you guys know what commands are needed.  If you could add them in, that would be most appreciated. 
Your users want to know just how easy it is. 

Your examples show how to start the quickstarters from the command line from your repository, but not from the web from frm my repository, and not from the web from your repository. 

I hope this helps.  We are almost there.  I think that this app store for end users is a huge untapped area.  All the docker hosting companies are focussed on the corporate market.  But they say that in general, the end user market is larger than the corporate market.